Begin working on dependency tree creation.
Basic idea: Components can have dependencies. If they do, those
dependency must be started before the component can be. A tree is built
from these dependencies, and then each level of the tree is started in
parallel, then the next level is started, and then the next level, and
so on.

When the dependency tree is being stopped, it is run through in reverse.